Sanitize file names
