+++ src/main/webapp/app/shared/monaco-editor/model/actions/attachment.action.ts
@@ -1,6 +1,7 @@
 import { faImage } from '@fortawesome/free-solid-svg-icons';
 import { TextEditorAction } from 'app/shared/monaco-editor/model/actions/text-editor-action.model';
 import { TextEditor } from 'app/shared/monaco-editor/model/actions/adapter/text-editor.interface';
+import { sanitizeStringForMarkdownEditor } from 'app/shared/util/markdown.util';
 
 interface AttachmentArguments {
     text: string;
@@ -34,7 +35,7 @@ export class AttachmentAction extends TextEditorAction {
         if (!args?.text || !args?.url) {
             this.replaceTextAtCurrentSelection(editor, AttachmentAction.DEFAULT_INSERT_TEXT);
         } else {
-            this.replaceTextAtCurrentSelection(editor, `![${args.text}](${args.url})`);
+            this.replaceTextAtCurrentSelection(editor, `![${sanitizeStringForMarkdownEditor(args.text)}](${args.url})`);
         }
         editor.focus();
     }
